optimize sidebar
