+++ src/components/Basic/BorrowTabs/RepayBorrowTab.js
@@ -282,7 +282,7 @@ function RepayBorrowTab({ asset, settings, changeTab, onCancel, setSetting }) {
               <span>Repay VAI Balance</span>
             </div>
             <span>
-              {getBigNumber(settings.vaiMinted)
+              {getBigNumber(settings.userVaiMinted)
                 .dp(2, 1)
                 .toString(10)}{' '}
               VAI
